no_comment_mouse

+++ ppr-tclsh/tclCmdMZ.c
@@ -170,7 +170,7 @@ Tcl_RegexpCmd(dummy, interp, argc, argv)
 	    }
 	} else {
 	    if (indices) {
-		snprintf(info, sizeof(info), "%d %d", start - string, end - string - 1);
+		snprintf(info, sizeof(info), "%d %d", (int)(start - string), (int)(end - string - 1));
 		result = Tcl_SetVar(interp, argPtr[i+2], info, 0);
 	    } else {
 		char savedChar, *first, *last;
@@ -1062,7 +1062,7 @@ Tcl_StringCmd(dummy, interp, argc, argv)
 		    " length string\"", (char *) NULL);
 	    return TCL_ERROR;
 	}
-	snprintf(interp->result, TCL_RESULT_SIZE+1, "%d", strlen(argv[2]));
+	snprintf(interp->result, TCL_RESULT_SIZE+1, "%d", (int)strlen(argv[2]));
 	return TCL_OK;
     } else if ((c == 'm') && (strncmp(argv[1], "match", length) == 0)) {
 	if (argc != 4) {
